pyproject.toml
[tool.poetry]
name = "netsource"
version = "0.1.0"
description = "A cross-platform application for generating live network radio programmes"
authors = ["Marc Steele <marc@steama.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kivy = "^2.3.0"
pyaudio = "^0.2.14"
pydantic = "^2.6.3"
numpy = "^1.26.4"
kivymd = "^1.2.0"
schedule = "^1.2.2"
python-dateutil = "^2.9.0.post0"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
